RAG/requirements.txt
fastapi
uvicorn
gradio
python-dotenv
pydantic
sentence-transformers
faiss-cpu==1.7.4
datasets
requests
numpy
torch
transformers